Switch to interface in transcript
